fixed arranger bug (arranger dependent on movementExecutor)
