--- Assets/Resources/graph2_small.json.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 73ab6108b39eb274fadd2287cbe644b9
-TextScriptImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
