+++ Assets/Scripts/GraphLoader.cs
@@ -27,7 +27,9 @@ namespace Assets.Scripts
     {
       //TODO: Try/catch file for bad formats.
       var graphJson = File.ReadAllText(filePath);
-      return JsonUtility.FromJson<Graph>(graphJson);
+      var graph = JsonUtility.FromJson<Graph>(graphJson);
+      Debug.Log($"Nodes count: {graph.nodes.Count}\nEdges count: {graph.edges.Count}");
+      return graph;
     }
   }
 }
